requirements.txt
# dev requirements
einops
numpy
torch
pillow